requirements.txt
asgiref==3.4.1
boto3==1.18.53
botocore==1.21.53
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.6
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
decouple==0.0.7
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.6
django-environ==0.6.0
django-heroku==0.3.1
django-storages==1.11.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
docutils==0.17.1
gunicorn==20.1.0
HttpPy==0.0.9
idna==3.2
itypes==1.2.0
Jinja2==3.0.2
jmespath==0.10.0
MarkupSafe==2.0.1
oauthlib==3.1.1
Pillow==8.3.2
psycopg2==2.9.1
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.3.0
python-dateutil==2.8.2
python-dotenv==0.19.1
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
s3transfer==0.5.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
uritemplate==4.1.1
urllib3==1.26.7
whitenoise==5.3.0
